% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSlide.R
\name{testing_1}
\alias{testing_1}
\title{Sum Function}
\usage{
testing_1(x, y, z)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{z}{Numeric vector.}
}
\value{
Numeric vector representing the sum of x, y, and z.
}
\description{
Adds three numeric vectors element-wise.
}
\examples{
testing_1(1, 2, 3)
testing_1(c(1, 2), c(3, 4), c(5, 6))
}
