% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSlide.R
\name{testing_2}
\alias{testing_2}
\title{Multiplication Function}
\usage{
testing_2(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Numeric vector representing the product of x and y.
}
\description{
Multiplies two numeric vectors element-wise.
}
\examples{
testing_2(2, 3)
testing_2(c(1, 2), c(3, 4))
}
