% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSlide.R
\name{window_binary}
\alias{window_binary}
\title{Determine direction of change in a window}
\usage{
window_binary(x)
}
\arguments{
\item{x}{Numeric vector representing one window of data.}
}
\value{
Numeric value: +1 or -1.
}
\description{
Returns +1 if the window shows an increase,
and -1 if it shows a decrease.
}
\examples{
window_binary(c(1, 2, 3))
window_binary(c(3, 2, 1))
}
