% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSlide.R
\name{window_mean}
\alias{window_mean}
\title{Compute mean value of a window}
\usage{
window_mean(x)
}
\arguments{
\item{x}{Numeric vector representing one window of data.}
}
\value{
Numeric value equal to the mean of the window.
}
\description{
Calculates the mean of a numeric vector representing a window.
}
\examples{
window_mean(c(1, 2, 3, 4))
window_mean(c(0.5, 0.7, 0.6))
}
